feat: setup frontend with TypeScript, Vite, and Bun
