+++ frontend/.bun-version
@@ -1,0 +1,1 @@
+1.1.38
