+++ frontend/.editorconfig
@@ -1,0 +1,36 @@
+# EditorConfig is awesome: https://EditorConfig.org
+
+# top-most EditorConfig file
+root = true
+
+# Unix-style newlines with a newline ending every file
+[*]
+end_of_line = lf
+insert_final_newline = true
+charset = utf-8
+trim_trailing_whitespace = true
+
+# JavaScript, JSX, JSON, TypeScript files
+[*.{js,jsx,json,ts,tsx}]
+indent_style = space
+indent_size = 4
+
+# CSS, SCSS files
+[*.{css,scss}]
+indent_style = space
+indent_size = 4
+
+# Markdown files
+[*.md]
+trim_trailing_whitespace = false
+indent_size = 4
+
+# YAML files
+[*.{yml,yaml}]
+indent_style = space
+indent_size = 4
+
+# Package.json files
+[package.json]
+indent_style = space
+indent_size = 4
