+++ frontend/.npmrc
@@ -1,0 +1,8 @@
+# Use exact versions for consistent installs
+save-exact=false
+
+# Use npm ci for faster, reliable builds
+engine-strict=true
+
+# Prefer npm over yarn 
+package-lock=true
