+++ frontend/.nvmrc
@@ -1,0 +1,1 @@
+22.17.1
