+++ frontend/.prettierignore
@@ -1,0 +1,39 @@
+# Dependencies
+node_modules
+package-lock.json
+pnpm-lock.yaml
+yarn.lock
+
+# Build outputs
+dist
+dist-ssr
+build
+.next
+out
+
+# Logs
+*.log
+
+# Environment files
+.env
+.env.local
+.env.*.local
+
+# Cache
+.cache
+.parcel-cache
+.vite
+
+# OS files
+.DS_Store
+Thumbs.db
+
+# IDE
+.idea
+.vscode
+*.swp
+*.swo
+
+# Generated files
+coverage
+*.tsbuildinfo
