+++ frontend/.prettierrc
@@ -1,0 +1,12 @@
+{
+    "semi": false,
+    "singleQuote": true,
+    "tabWidth": 4,
+    "trailingComma": "es5",
+    "printWidth": 80,
+    "arrowParens": "avoid",
+    "endOfLine": "lf",
+    "bracketSpacing": true,
+    "jsxSingleQuote": false,
+    "plugins": []
+}
