+++ frontend/.vscode/extensions.json
@@ -1,0 +1,11 @@
+{
+  "recommendations": [
+    "esbenp.prettier-vscode", 
+    "dbaeumer.vscode-eslint",
+    "bradlc.vscode-tailwindcss",
+    "ms-vscode.vscode-typescript-next",
+    "usernamehw.errorlens",
+    "formulahendry.auto-rename-tag",
+    "christian-kohler.path-intellisense"
+  ]
+}
