--- frontend/src/App.jsx
@@ -1,26 +1,0 @@
-import { useState } from 'react'
-
-function App() {
-  const [count, setCount] = useState(0)
-
-  return (
-    <div className="min-h-screen bg-gray-100 flex items-center justify-center">
-      <div className="bg-white p-8 rounded-lg shadow-lg text-center">
-        <h1 className="text-3xl font-bold text-gray-800 mb-4">
-          React + Tailwind CSS
-        </h1>
-        <p className="text-gray-600 mb-6">
-          Your frontend is ready!
-        </p>
-        <button
-          onClick={() => setCount((count) => count + 1)}
-          className="bg-blue-500 hover:bg-blue-600 text-white font-semibold py-2 px-4 rounded transition-colors"
-        >
-          Count: {count}
-        </button>
-      </div>
-    </div>
-  )
-}
-
-export default App
