+++ frontend/src/index.css
@@ -1,1 +1,1 @@
-@import "tailwindcss";
+@import 'tailwindcss';
